+++ src/Aspire.Dashboard/Components/Dialogs/SettingsDialog.razor.cs
@@ -83,7 +83,7 @@ public partial class SettingsDialog : IDialogContentComponent, IAsyncDisposable
         }
         catch (JSDisconnectedException)
         {
-            // Per https://learn.microsoft.com/en-us/aspnet/core/blazor/javascript-interoperability/?view=aspnetcore-7.0#javascript-interop-calls-without-a-circuit
+            // Per https://learn.microsoft.com/aspnet/core/blazor/javascript-interoperability/?view=aspnetcore-7.0#javascript-interop-calls-without-a-circuit
             // this is one of the calls that will fail if the circuit is disconnected, and we just need to catch the exception so it doesn't pollute the logs
         }
     }
