+++ src/Aspire.Hosting.Azure.Provisioning/JsonExtensions.cs
@@ -1,5 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
+
 using System.Text.Json.Nodes;
 
 namespace Aspire.Hosting.Azure;
