+++ src/Components/Aspire.Azure.Security.KeyVault/AspireKeyVaultExtensions.cs
@@ -16,6 +16,9 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 
 namespace Microsoft.Extensions.Hosting;
 
+/// <summary>
+/// Provides extension methods for registering and configuring Azure Key Vault secrets in a .NET Aspire application.
+/// </summary>
 public static class AspireKeyVaultExtensions
 {
     internal const string DefaultConfigSectionName = "Aspire:Azure:Security:KeyVault";
