+++ src/Microsoft.Extensions.ServiceDiscovery.Abstractions/Features/IEndPointLoadFeature.cs
@@ -3,9 +3,14 @@
 
 namespace Microsoft.Extensions.ServiceDiscovery.Abstractions;
 
+/// <summary>
+/// Represents a feature that provides information about the current load of an endpoint.
+/// </summary>
 public interface IEndPointLoadFeature
 {
-    // CurrentLoad is some comparable measure of load (queue length, concurrent requests, etc)
+    /// <summary>
+    /// Gets a comparable measure of the current load of the endpoint (e.g. queue length, concurrent requests, etc).
+    /// </summary>
     public double CurrentLoad { get; }
 }
